+++ src/components/util/storybook.js
@@ -1,3 +1,7 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import FDS from '../../../lib/dictionary/js/styleConstants';
+
 /**
  * Storybook helpers
  */
@@ -26,4 +30,30 @@ export const storyBackgrounds = {
     'url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAJ0lEQVQYV2NkQAP/b///z4gsBhZQZWSEC8IEQIrAgsgCYEF0AZAgAAvKE968p7/mAAAAAElFTkSuQmCC)',
   check:
     'url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAcAAAAHCAYAAADEUlfTAAAAHElEQVQYV2M8fPiwLwMOwAiStLW13YxNftBJAgAx2BqeI9XcBAAAAABJRU5ErkJggg==)',
+};
+
+export const StoryWrapper = ({ children }) => (
+  <div
+    className="display--inlineFlex"
+    style={{
+      flexWrap: 'wrap',
+      marginBottom: `-${FDS.SPACE_HALF}`,
+      marginRight: `-${FDS.SPACE_HALF}`,
+      alignItems: 'center',
+    }}
+  >
+    {children}
+  </div>
+);
+
+StoryWrapper.propTypes = {
+  children: PropTypes.node,
+};
+
+export const StoryItem = ({ children }) => (
+  <div className="margin--right--half margin--bottom--half">{children}</div>
+);
+
+StoryItem.propTypes = {
+  children: PropTypes.node,
 };